+++ awesome-core/src/main/java/com/awesome/configurations/JPAConfig.java
@@ -1,16 +1,14 @@
 package com.awesome.configurations;
 
-import com.awesome.domains.Project.entities.ProjectEntities;
-import com.awesome.domains.ProjectTask.entities.ProjectTaskEntities;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @Configuration
-@EntityScan(basePackageClasses = {ProjectEntities.class, ProjectTaskEntities.class})
+@EntityScan(basePackages = {"com.awesome.domains"})
 @EnableTransactionManagement
-@EnableJpaRepositories(basePackageClasses = {ProjectEntities.class, ProjectTaskEntities.class})
+@EnableJpaRepositories(basePackages = {"com.awesome.domains"})
 public class JPAConfig {
 
 }
