--- awesome-core/src/main/java/com/awesome/domains/Project/ProjectRepository.java
@@ -1,5 +1,0 @@
-package com.awesome.domains.Project;
-
-// Project의 레포지토리
-public class ProjectRepository {
-}
