--- awesome-core/src/main/java/com/awesome/domains/Project/entities/ProjectDAO.java
@@ -1,9 +1,0 @@
-package com.awesome.domains.Project.entities;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ProjectDAO extends JpaRepository<ProjectEntity, Long> {
-  List<ProjectEntity> findAllByProjectNameLike(String projectName);
-}
