--- awesome-core/src/main/java/com/awesome/domains/Project/entities/ProjectEntities.java
@@ -1,4 +1,0 @@
-package com.awesome.domains.Project.entities;
-
-public interface ProjectEntities {
-}
