--- awesome-core/src/main/java/com/awesome/domains/Project/enums/TaskType.java
@@ -1,11 +1,0 @@
-package com.awesome.domains.Project.enums;
-
-/**
- * Scope
- */
-public enum TaskType {
-    //Task 에 발생한 이슈 (부모는 타스크여야함)
-    ISSUE,
-    //Project의 일
-    TASK;
-}
