--- awesome-core/src/main/java/com/awesome/domains/ProjectTask/ProjectTaskRepository.java
@@ -1,5 +1,0 @@
-package com.awesome.domains.ProjectTask;
-
-// ProjectTask의 레포지토리
-public class ProjectTaskRepository {
-}
