--- awesome-core/src/main/java/com/awesome/domains/ProjectTask/entities/ProjectTaskEntities.java
@@ -1,4 +1,0 @@
-package com.awesome.domains.ProjectTask.entities;
-
-public interface ProjectTaskEntities {
-}
