+++ awesome-core/src/main/java/com/awesome/domains/project/enums/ProjectStatus.java
@@ -1,0 +1,17 @@
+package com.awesome.domains.project.enums;
+
+/**
+ * 프로젝트 상태
+ */
+public enum ProjectStatus {
+    // 앞으로 해야함
+    TODO,
+    // 일시 중단됨
+    STOPPED,
+    // 진행중
+    PROGRESS,
+    // 취소됨
+    CANCELED,
+    // 완료됨
+    RESOLVED
+}
