+++ awesome-core/src/main/java/com/awesome/domains/projecttask/enums/TaskType.java
@@ -1,0 +1,11 @@
+package com.awesome.domains.projecttask.enums;
+
+/**
+ * Scope
+ */
+public enum TaskType {
+    //Task 에 발생한 이슈 (부모는 타스크여야함)
+    ISSUE,
+    //Project의 일
+    TASK;
+}
