+++ awesome-core/src/main/java/com/awesome/domains/user/entities/UserDAO.java
@@ -1,0 +1,9 @@
+package com.awesome.domains.user.entities;
+
+import com.awesome.domains.project.entities.ProjectEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface UserDAO extends JpaRepository<UserEntity, Long> {
+}
