+++ awesome-core/src/main/java/com/awesome/domains/user/enums/UserPosition.java
@@ -1,0 +1,8 @@
+package com.awesome.domains.user.enums;
+
+public enum UserPosition {
+    // 팀원
+    TEAMMEMBER,
+    // 팀장
+    TEAMLEADER
+}
