+++ awesome-core/src/test/java/com/awesome/domains/project/entities/AwesomeBootTest.java
@@ -1,0 +1,15 @@
+package com.awesome.domains.project.entities;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.TYPE)
+@Retention(
+        RetentionPolicy.RUNTIME)
+@Inherited
+@ActiveProfiles(value = {"local"})
+@SpringBootTest(properties = "spring.profiles.active:local")
+public @interface AwesomeBootTest {
+}
