+++ awesome-core/src/test/java/com/awesome/domains/project/services/ProjectTXServiceTest.java
@@ -1,0 +1,21 @@
+package com.awesome.domains.project.services;
+
+import com.awesome.applications.tx.ProjectTXService;
+import com.awesome.domains.project.entities.AwesomeBootTest;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@AwesomeBootTest
+class ProjectTXServiceTest {
+
+    @Autowired
+    ProjectTXService projectTXService;
+
+    @Test
+    void testServiceNotNull() {
+
+        assertNotNull(projectTXService);
+    }
+}
